[feat] 그려짐

+++ src/string/ft_stof.c
@@ -4,7 +4,7 @@
 
 float	ft_stof(char const *str)
 {
-	double	ret;
+	float	ret;
 	int		digit;
 	int		digit_decimal;
 	int		sign;
